refactor: main page icon, timetable image files 변경 - #206
